refactor(admin): extract storage policy logic into custom hooks
Split large AdminPoliciesPage component into focused custom hooks and utilities to improve maintainability and testability:

- Add policyActionSelection module with tests for determining connection test modes and value sources based on policy state
- Add policyFormComparison utility for detecting unsaved changes with proper normalization
- Add policyFormTransition module with tests for handling driver type changes and form field updates
- Extract useStoragePolicyActionController for managing policy actions (connection tests, CORS config, driver promotion, auth flows)
- Extract useStoragePolicyDescriptorController for remote node and driver descriptor management
- Extract useStoragePolicyEditorController for policy creation/update workflows and wizard step navigation
- Extract useStoragePolicyListController for pagination, sorting, and deletion
- Extract useStoragePolicyMigrationController for migration dry-run and execution
- Reduce AdminPoliciesPage from 1595 to 235 lines by delegating logic to specialized hooks
